Fixed gurobi bug
This is a small thing I just noticed. In the solver.jl file in the examples, you have optimizers_list = ["cplex", "cbc", "ipopt", "juniper", "alpine", "glpk"] and gurobi is not in this. This results in an error when using gurobi (ERROR: LoadError: UndefVarError: Memento not defined). If you just add "gurobi" to the beginning of that list it works (since it an option in the if/else branching later).